feature: rename abstract table classes to repositories
